requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pandas>=2.1.0
requests>=2.31.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23
PyPDF2>=3.0.0
pdfplumber>=0.10.0
PyYAML>=6.0.1
prometheus-client>=0.19.0
